add connect function to landing page
